refactor: rename OverlayPage to OverlayWindow  to ensure consistency in the naming of the main components
