+++ src/main.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { HashRouter, Routes, Route } from 'react-router-dom';
 import { MainWindow } from './pages/MainWindow/MainWindow';
-import { OverlayPage } from './pages/OverlayPage/OverlayPage';
+import { OverlayWindow } from './pages/OverlayPage/OverlayWindow';
 import './i18n';
 import './styles/index.scss';
 
@@ -11,7 +11,7 @@ ReactDOM.createRoot(document.getElementById('root')!).render(
     <HashRouter>
       <Routes>
         <Route path="/" element={<MainWindow />} />
-        <Route path="/overlay" element={<OverlayPage />} />
+        <Route path="/overlay" element={<OverlayWindow />} />
       </Routes>
     </HashRouter>
   </React.StrictMode>
