+++ src/pages/OverlayPage/OverlayWindow.tsx
@@ -1,10 +1,10 @@
 import { useEffect } from 'react';
 import { getCurrentWebviewWindow } from '@tauri-apps/api/webviewWindow';
 import { telemetryConnectionStore } from '../../store/iracing/telemetry-connection.store';
-import { OverlayCanvas } from '../../components/OverlayCanvas/OverlayCanvas';
+import { OverlayCanvas } from './OverlayCanvas/OverlayCanvas';
 import { initOverlaySync } from '../../store/sync/sync-init';
 
-export const OverlayPage = () => {
+export const OverlayWindow = () => {
   useEffect(() => {
     void telemetryConnectionStore.startWidgetListener();
